Removed support for systems without SSE, which didn't actually exist but there was a #define for it. Improved SSE-based vector classes, organized them into a basic "math library". Fixed a bug (in theory?) where SSE vectors were being directly cast to/from memory addresses. Apparently you're supposed to use the load and store functions to do this, but it seemed to work anyway... fixed now though. Added endian-swapping functions. Added support for compile-time-constant bool type. Added string concatenation operator to the language. Improved compiler path-matching. Previously it used string comparison, now it queries the OS to determine if two string paths point to the same resource on disk. Task memory query now takes task function context as its input instead of just the constant arguments so that it can access fields like sample rate. Added another optional function to tasks: the task initializer. This function runs after the memory query and is used to initialize the task memory. It is also used to submit sample requests (see below). Added a class c_sample_library which manages samples loaded from disk (e.g. wavs). Samples are requested upon synth initialization, the list of requested samples is diffed against the list of currently loaded samples, and samples are unloaded/loaded/reloaded as needed. Currently wav is the only format supported. Samples can be loaded as looping or bidi-looping, and are preprocessed so that they can be efficiently played with these settings. (Note that loops will be driven by cue points in the sample files, but reading cue points is not yet implemented.) Added a set of "native samples" for basic waveforms including sine, triangle, and sawtooth, which are always loaded. These are generated offline and stored as constant arrays (the code to do this is included but #ifdef'd out). Native samples are built as mipmaps in order to avoid frequencies above nyquist when resampling/pitch-bending. Wrote part of the sampler native module. The sampler takes a string name, channel, loop settings, and (variable) speed, and will play back a sample according to these settings. All logic, including mipmap selection/blending, has been implemented except for the sampling itself (all samples are returned as 0 currently), however this has not been tested yet.

+++ WaveLang/source/runtime/runtime_context.cpp
@@ -5,7 +5,7 @@ void s_runtime_context::stream_callback(const s_driver_stream_callback_context &
 	s_runtime_context *this_ptr = static_cast<s_runtime_context *>(context.user_data);
 
 	s_executor_chunk_context chunk_context;
-	chunk_context.sample_rate = context.driver_settings->sample_rate;
+	chunk_context.sample_rate = static_cast<uint32>(context.driver_settings->sample_rate);
 	chunk_context.output_channels = context.driver_settings->output_channels;
 	chunk_context.sample_format = context.driver_settings->sample_format;
 	chunk_context.frames = context.driver_settings->frames_per_buffer;
